feat: implement ScrollToTop component and refactor scroll handling in multiple pages
