+++ src/components/Pages/PageArticle.tsx
@@ -1,12 +1,17 @@
 'use client'
 
-import { AnimatePresence, motion } from 'framer-motion'
-import { ArrowLeft, ChevronDown } from 'lucide-react'
+import { ArrowLeft } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import React from 'react'
 import type { Dictionary } from '@/app/dictionaries'
-import { Button, SectionHero, ShareSocial, Spinner } from '@/components'
+import {
+	Button,
+	ScrollToTop,
+	SectionHero,
+	ShareSocial,
+	Spinner,
+} from '@/components'
 import type { Article } from '@/types'
 import { formatDate } from '@/utils'
 
@@ -20,21 +25,6 @@ function PageArticle({ dict, language, id }: Readonly<PageArticleProps>) {
 	const [article, setArticle] = React.useState<Article | null>()
 	const [isPending, startTransition] = React.useTransition()
 	const router = useRouter()
-
-	// Stato per il tasto Scroll to Top
-	const [showScrollTop, setShowScrollTop] = React.useState<boolean>(false)
-
-	React.useEffect(() => {
-		const handleScrollWindow = () => {
-			setShowScrollTop(window.scrollY > 400)
-		}
-		window.addEventListener('scroll', handleScrollWindow)
-		return () => window.removeEventListener('scroll', handleScrollWindow)
-	}, [])
-
-	const scrollToTop = () => {
-		window.scrollTo({ top: 0, behavior: 'smooth' })
-	}
 
 	React.useEffect(() => {
 		startTransition(() => {
@@ -72,23 +62,7 @@ function PageArticle({ dict, language, id }: Readonly<PageArticleProps>) {
 					title={article.title}
 				>
 					{/* BOTTONE SCROLL TO TOP */}
-					<AnimatePresence>
-						{showScrollTop && (
-							<div className="fixed right-6 bottom-36 z-50 flex">
-								<motion.button
-									initial={{ opacity: 0, scale: 0 }}
-									animate={{ opacity: 1, scale: 1 }}
-									exit={{ opacity: 0, scale: 0 }}
-									whileHover={{ scale: 1.1 }}
-									whileTap={{ scale: 0.9 }}
-									onClick={scrollToTop}
-									className="flex h-12 w-12 cursor-pointer items-center justify-center rounded-full bg-white shadow-2xl ring-1 ring-black/5 transition-colors hover:bg-slate-50 md:h-14 md:w-14 dark:bg-slate-800 dark:ring-white/10 dark:hover:bg-slate-700"
-								>
-									<ChevronDown className="h-6 w-6 rotate-180 text-primary md:h-7 md:w-7" />
-								</motion.button>
-							</div>
-						)}
-					</AnimatePresence>
+					<ScrollToTop />
 
 					<div className="flex flex-col gap-2">
 						<div className="flex flex-col gap-2 text-black md:flex-row dark:text-white">
