+++ src/components/Pages/PageBlog.tsx
@@ -10,11 +10,13 @@ import {
 	Button,
 	Input,
 	Label,
+	ScrollToTop,
 	SectionHero,
 	Select,
 	Skeleton,
 } from '@/components'
 import { CardArticle, FeaturedArticle } from '@/components/UI/Cards'
+import { useScrollTop } from '@/hooks'
 import type { Article } from '@/types'
 
 interface PageBlogProps {
@@ -36,17 +38,8 @@ function PageBlog({ dict, language }: Readonly<PageBlogProps>) {
 	// 👁️‍🗨️ Toggle filtri
 	const [showFilters, setShowFilters] = React.useState(false)
 
-	// 🖱️ Gestione Scroll Pulsante
-	const [isScrolled, setIsScrolled] = React.useState(false)
-
-	React.useEffect(() => {
-		const handleScroll = () => {
-			// Soglia più bassa per evitare che sparisca sopra
-			setIsScrolled(window.scrollY > 200)
-		}
-		window.addEventListener('scroll', handleScroll)
-		return () => window.removeEventListener('scroll', handleScroll)
-	}, [])
+	// 🖱️ Gestione Scroll Pulsante (Soglia più bassa per evitare che sparisca sopra)
+	const { showScrollTop: isScrolled } = useScrollTop(200)
 
 	// 🔄 Funzione fetch con filtri
 	const fetchArticlesWithFilters = React.useCallback(() => {
@@ -164,22 +157,7 @@ function PageBlog({ dict, language }: Readonly<PageBlogProps>) {
 				)}
 
 				{/* 🖱️ Scroll Up Button (Fixed, appare quando si scende o se non c'è FeaturedArticle) */}
-				<AnimatePresence>
-					{isScrolled && (
-						<div className="fixed right-6 bottom-36 z-50 flex">
-							<motion.button
-								layoutId="scroll-button"
-								initial={{ opacity: 0, scale: 0, rotate: 0 }}
-								animate={{ opacity: 1, scale: 1, rotate: 180 }}
-								exit={{ opacity: 0, scale: 0, rotate: 0 }}
-								onClick={() => window.scrollTo({ top: 0, behavior: 'smooth' })}
-								className="flex h-12 w-12 cursor-pointer items-center justify-center rounded-full bg-white shadow-2xl ring-1 ring-black/5 transition-colors hover:bg-slate-50 md:h-14 md:w-14 dark:bg-slate-800 dark:ring-white/10 dark:hover:bg-slate-700"
-							>
-								<ChevronDown className="h-6 w-6 text-primary md:h-7 md:w-7" />
-							</motion.button>
-						</div>
-					)}
-				</AnimatePresence>
+				<ScrollToTop />
 
 				{/* 🔍 SEZIONE FILTRI */}
 				<div id="articles-section" className="flex scroll-mt-24 flex-col gap-6">
