+++ src/components/Pages/PageSkills.tsx
@@ -1,8 +1,7 @@
 'use client'
 
-import { AnimatePresence, motion } from 'framer-motion'
 import * as LucideIcons from 'lucide-react'
-import { ChevronDown, Wrench } from 'lucide-react'
+import { Wrench } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import React from 'react'
 import type { Dictionary } from '@/app/dictionaries'
@@ -13,6 +12,7 @@ import {
 	CardContent,
 	Dialog,
 	ProgressBar,
+	ScrollToTop,
 	SectionHero,
 	Tabs,
 } from '@/components'
@@ -37,21 +37,6 @@ export default function PageSkills({
 	const hardSkills = Object.keys(skills).filter((key) => key !== 'soft')
 	const softSkills = Object.keys(skills).filter((key) => key === 'soft')
 	const [showLegend, setShowLegend] = React.useState<boolean>(false)
-
-	// Stato per il tasto Scroll to Top
-	const [showScrollTop, setShowScrollTop] = React.useState<boolean>(false)
-
-	React.useEffect(() => {
-		const handleScrollWindow = () => {
-			setShowScrollTop(window.scrollY > 400)
-		}
-		window.addEventListener('scroll', handleScrollWindow)
-		return () => window.removeEventListener('scroll', handleScrollWindow)
-	}, [])
-
-	const scrollToTop = () => {
-		window.scrollTo({ top: 0, behavior: 'smooth' })
-	}
 
 	const mappa = {
 		0: dict.skills.languages,
@@ -114,23 +99,7 @@ export default function PageSkills({
 			subtitle={dict.skills.subtitle}
 		>
 			{/* BOTTONE SCROLL TO TOP */}
-			<AnimatePresence>
-				{showScrollTop && (
-					<div className="fixed right-6 bottom-36 z-50 flex">
-						<motion.button
-							initial={{ opacity: 0, scale: 0 }}
-							animate={{ opacity: 1, scale: 1 }}
-							exit={{ opacity: 0, scale: 0 }}
-							whileHover={{ scale: 1.1 }}
-							whileTap={{ scale: 0.9 }}
-							onClick={scrollToTop}
-							className="flex h-12 w-12 cursor-pointer items-center justify-center rounded-full bg-white shadow-2xl ring-1 ring-black/5 transition-colors hover:bg-slate-50 md:h-14 md:w-14 dark:bg-slate-800 dark:ring-white/10 dark:hover:bg-slate-700"
-						>
-							<ChevronDown className="h-6 w-6 rotate-180 text-primary md:h-7 md:w-7" />
-						</motion.button>
-					</div>
-				)}
-			</AnimatePresence>
+			<ScrollToTop />
 
 			<div className="flex flex-col justify-center px-2 md:px-14">
 				<div className="block">
