+++ src/components/UI/index.ts
@@ -20,6 +20,7 @@ export { default as LoadingScreen } from './LoadingScreen'
 export { default as ModalMessage } from './ModalMessage'
 export { default as ProgressBar } from './ProgressBar'
 export { default as QuizResultPopup } from './QuizResultPopup'
+export { default as ScrollToTop } from './ScrollToTop'
 export { default as SectionHero } from './SectionHero'
 export { default as Select } from './Select'
 export { default as ServiceCard } from './ServiceCard'
