+++ src/hooks/index.ts
@@ -1,1 +1,3 @@
 export { default as useScreenDimensions } from './useScreenDimensions'
+export { default as useScrollReveal } from './useScrollReveal'
+export { default as useScrollTop } from './useScrollTop'
